I refactored the public headers to be inside a pharovm subfolder.
