+++ src/aioWin.c
@@ -1,4 +1,4 @@
-#include "pharo.h"
+#include "pharovm/pharo.h"
 #include "aioWin.h"
 
 sqInt ioRelinquishProcessorForMicroseconds(sqInt microSeconds){
