+++ src/fileDialogWin32.c
@@ -4,8 +4,8 @@
 #define UNICODE
 #endif
 
-#include "fileDialog.h"
-#include "stringUtilities.h"
+#include "pharovm/fileDialog.h"
+#include "pharovm/stringUtilities.h"
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 #include <Shobjidl.h>
