+++ src/parameterVector.c
@@ -1,4 +1,4 @@
-#include "parameterVector.h"
+#include "pharovm/parameterVector.h"
 #include <stdlib.h>
 
 VMErrorCode
