+++ src/win32Main.c
@@ -1,4 +1,4 @@
-#include "pharoClient.h"
+#include "pharovm/pharoClient.h"
 
 #define WIN32_LEAN_AND_MEAN
 #include <windows.h>
